refactor(core): use_cases to change alarm status
